requirements.txt
Django
django-import-export
django-simpleui
simplepro
Pillow
requests